Avoid libmpv being deployed for client
DedupEngine and VideoFile is now in a separate project that is only
used by the server/console application. While the VideoDedupShared
project has another VideoFile class to provide functionality for WCF
and client. Both a compatible due to the IVideoFile interface.
