+++ DedupEngine/MpvLib/DataFormat.cs
@@ -1,0 +1,16 @@
+namespace DedupEngine.MpvLib
+{
+    internal enum DataFormat
+    {
+        None = 0,
+        String = 1,
+        OsdString = 2,
+        Flag = 3,
+        Int64 = 4,
+        Double = 5,
+        Node = 6,
+        NodeArray = 7,
+        NodeMap = 8,
+        ByteArray = 9
+    }
+}
