+++ DedupEngine/MpvLib/Event.cs
@@ -1,0 +1,14 @@
+namespace DedupEngine.MpvLib
+{
+    using System;
+    using System.Runtime.InteropServices;
+
+    [StructLayout(LayoutKind.Sequential)]
+    internal class Event
+    {
+        public int Id;
+        public int Error;
+        public ulong UserData;
+        public IntPtr Data;
+    }
+}
