+++ DedupEngine/MpvLib/MpvException.cs
@@ -1,0 +1,10 @@
+namespace DedupEngine.MpvLib
+{
+    using System;
+
+    public class MpvException : Exception
+    {
+        public MpvException() : base() { }
+        public MpvException(string message) : base(message) { }
+    }
+}
