+++ VideoDedup.sln
@@ -17,6 +17,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VideoDedupConsole", "VideoD
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SmartTimer", "HandyTimer\SmartTimer.csproj", "{BA276BCA-E923-4E5C-A483-AC04DDF0EB61}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DedupEngine", "DedupEngine\DedupEngine.csproj", "{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -73,6 +75,14 @@ Global
 		{BA276BCA-E923-4E5C-A483-AC04DDF0EB61}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BA276BCA-E923-4E5C-A483-AC04DDF0EB61}.Release|x86.ActiveCfg = Release|Any CPU
 		{BA276BCA-E923-4E5C-A483-AC04DDF0EB61}.Release|x86.Build.0 = Release|Any CPU
+		{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}.Debug|x86.Build.0 = Debug|Any CPU
+		{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}.Release|x86.ActiveCfg = Release|Any CPU
+		{6931A4E9-4586-44EB-92FC-8DC0DCDF0CA8}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
