+++ VideoDedupShared/Classes/VideoFile.cs
@@ -1,175 +1,61 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("System.Runtime.Serialization")]
-[assembly: InternalsVisibleTo("Newtonsoft.Json")]
 namespace VideoDedupShared
 {
     using System;
     using System.Collections.Generic;
     using System.Configuration;
-    using System.Diagnostics;
     using System.Drawing;
     using System.IO;
     using System.Linq;
     using System.Runtime.Serialization;
     using System.Threading;
-    using Newtonsoft.Json;
-    using VideoDedupShared.MpvLib;
-    using XnaFan.ImageComparison;
 
     [DataContract]
     public class VideoFile : IVideoFile, IDisposable
     {
         internal VideoFile() { }
 
-        public VideoFile(
-            VideoFile other,
-            int imageCount = 0)
+        public VideoFile(IVideoFile other)
         {
             if (other is null)
             {
                 throw new ArgumentNullException(nameof(other));
             }
 
-            ImageCount = imageCount;
-
-            // We have to do a bit more work here since we can't use
-            // the other ctors because we don't want to eagerly load
-            // the images when we can copy them from the other file.
-            filePath = other.filePath;
-            fileSize = other.fileSize;
-            duration = other.duration;
-            codecInfo = other.codecInfo;
-
-            if (ImageCount == 0)
-            {
-                return;
-            }
-
-            if (other.ImageCount == ImageCount)
-            {
-                imageStreams = other.imageStreams
-                    .Select(image =>
-                    {
-                        var ms = new MemoryStream();
-                        image.Position = 0;
-                        image.CopyTo(ms);
-                        return ms;
-                    })
-                    .ToList();
-                return;
-            }
-
-            using (var mpv = new MpvWrapper(FilePath, imageCount, Duration))
-            {
-                imageStreams = mpv.GetImages(0, imageCount).ToList();
-            }
-        }
-
-        public VideoFile(
-            IVideoFile other,
-            int imageCount = 0)
-            : this(other.FilePath, imageCount)
-        {
-            if (other is null)
-            {
-                throw new ArgumentNullException(nameof(other));
-            }
-
+            filePath = other.FilePath;
             fileSize = other.FileSize;
             duration = other.Duration;
             codecInfo = other.CodecInfo;
+
+            imageStreams = other.ImageStreams.Select(ms =>
+            {
+                ms.Position = 0;
+                var @new = new MemoryStream();
+                ms.CopyTo(@new);
+                return @new;
+            }).ToList();
         }
 
-        public VideoFile(
-            string filePath,
-            int imageCount = 0)
-        {
-            if (string.IsNullOrWhiteSpace(filePath))
-            {
-                throw new ArgumentException($"'{nameof(filePath)}' cannot be" +
-                    $" null or whitespace", nameof(filePath));
-            }
-
-            this.filePath = filePath;
-            ImageCount = imageCount;
-
-            if (ImageCount == 0)
-            {
-                return;
-            }
-
-            using (var mpv = new MpvWrapper(FilePath, ImageCount, Duration))
-            {
-                imageStreams = mpv.GetImages(0, ImageCount).ToList();
-            }
-        }
-
-        [JsonIgnore]
         public string FilePath => filePath;
         [DataMember]
-        [JsonProperty]
         protected internal string filePath;
 
-        [JsonIgnore]
         public string FileName => Path.GetFileName(FilePath);
 
-        [JsonIgnore]
-        public long FileSize
-        {
-            get
-            {
-                if (!fileSize.HasValue)
-                {
-                    fileSize = new FileInfo(FilePath).Length;
-                }
-                return fileSize.Value;
-            }
-        }
-        [JsonProperty]
+        public long FileSize => fileSize;
         [DataMember]
-        protected internal long? fileSize = null;
+        protected internal long fileSize;
 
-        [JsonIgnore]
-        public TimeSpan Duration
-        {
-            get
-            {
-                if (duration == TimeSpan.Zero)
-                {
-                    try
-                    {
-                        duration = MpvWrapper.GetDuration(filePath);
-                    }
-                    catch (Exception)
-                    {
-                        duration = TimeSpan.Zero;
-                    }
-                }
-                return duration;
-            }
-        }
-        [JsonProperty]
+        public TimeSpan Duration => duration;
         [DataMember]
-        protected internal TimeSpan duration = TimeSpan.Zero;
+        protected internal TimeSpan duration;
 
-        [JsonIgnore]
-        public CodecInfo CodecInfo
-        {
-            get
-            {
-                if (codecInfo == null)
-                {
-                    codecInfo = MpvWrapper.GetCodecInfo(filePath);
-                }
-                return codecInfo;
-            }
-        }
-        [JsonIgnore]
+        public CodecInfo CodecInfo => codecInfo;
         [DataMember]
-        protected internal CodecInfo codecInfo = null;
+        protected internal CodecInfo codecInfo;
 
-        [JsonIgnore]
         public IEnumerable<Image> Images
         {
             get
@@ -181,115 +67,10 @@ namespace VideoDedupShared
                 return imageStreams.Select(image => Image.FromStream(image));
             }
         }
-        [JsonIgnore]
+        public IEnumerable<MemoryStream> ImageStreams => imageStreams;
         [DataMember]
         protected internal IList<MemoryStream> imageStreams =
             new List<MemoryStream>();
-
-        [JsonIgnore]
-        private int ImageCount { get; set; }
-
-        public bool IsDurationEqual(IVideoFile other,
-            IDurationComparisonSettings settings)
-        {
-            var diff = Math.Abs((Duration - other.Duration).TotalSeconds);
-            switch (settings.DifferenceType)
-            {
-                case DurationDifferenceType.Seconds:
-                    return diff < settings.MaxDifferenceSeconds;
-                case DurationDifferenceType.Percent:
-                    var max_diff = Duration.TotalSeconds / 100
-                        * settings.MaxDifferencePercent;
-                    return diff < max_diff;
-                default:
-                    throw new ConfigurationErrorsException(
-                        "DurationDifferenceType has not valid value");
-            }
-        }
-
-        protected MemoryStream GetImage(int index,
-            int imageCount, IImageComparisonSettings settings)
-        {
-            if (index >= imageCount || index < 0)
-            {
-                throw new ArgumentOutOfRangeException(nameof(index),
-                    "Index out of range.");
-            }
-
-            if (ImageCount != 0 && ImageCount != imageCount)
-            {
-                DisposeImages();
-            }
-            ImageCount = imageCount;
-
-            if (index < imageStreams.Count())
-            {
-                return imageStreams[index];
-            }
-
-            using (var mpv = new MpvWrapper(
-                filePath,
-                imageCount,
-                Duration))
-            {
-                // If its the first image we load, we already load
-                // the minimum number of images we need to compare.
-                // If it's not the first one, then we load all of them
-                // because we most likely need them.
-                var imagesToLoad = imageCount - index;
-                if (index == 0 && imageCount >= settings.MaxDifferentImages + 1)
-                {
-                    imagesToLoad = settings.MaxDifferentImages + 1;
-                }
-                foreach (var image in mpv.GetImages(index, imagesToLoad))
-                {
-                    imageStreams.Add(image);
-                }
-                return imageStreams[index];
-            }
-        }
-
-        public bool AreImagesEqual(VideoFile other,
-            IImageComparisonSettings settings,
-            CancellationToken cancelToken)
-        {
-            if (other is null)
-            {
-                throw new ArgumentNullException(nameof(other));
-            }
-
-            if (settings is null)
-            {
-                throw new ArgumentNullException(nameof(settings));
-            }
-
-            var differernceCount = 0;
-            foreach (var index in Enumerable.Range(0, settings.MaxCompares))
-            {
-                if (cancelToken.IsCancellationRequested)
-                {
-                    return false;
-                }
-
-                var thisImageSample =
-                    Image.FromStream(GetImage(index, settings.MaxCompares, settings));
-                var otherImageSample =
-                    Image.FromStream(other.GetImage(index, settings.MaxCompares, settings));
-
-                var diff = thisImageSample.PercentageDifference(otherImageSample);
-
-                if (diff > (double)settings.MaxDifferencePercent / 100)
-                {
-                    ++differernceCount;
-                }
-
-                if (differernceCount > settings.MaxDifferentImages)
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
 
         public override bool Equals(object obj) => Equals(obj as IVideoFile);
