--- VideoDedupShared/MpvLib/MpvException.cs
@@ -1,10 +1,0 @@
-namespace VideoDedupShared.MpvLib
-{
-    using System;
-
-    public class MpvException : Exception
-    {
-        public MpvException() : base() { }
-        public MpvException(string message) : base(message) { }
-    }
-}
