+++ VideoDedupShared/VideoDedupShared.csproj
@@ -32,12 +32,6 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Mpv.NET, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Mpv.NET.1.1.1\lib\netstandard2.0\Mpv.NET.dll</HintPath>
-    </Reference>
-    <Reference Include="Newtonsoft.Json, Version=12.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
-      <HintPath>..\packages\Newtonsoft.Json.12.0.3\lib\net45\Newtonsoft.Json.dll</HintPath>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
@@ -52,14 +46,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="MpvLib\DataFormat.cs" />
-    <Compile Include="MpvLib\Event.cs" />
-    <Compile Include="MpvLib\EventId.cs" />
-    <Compile Include="MpvLib\MpvException.cs" />
-    <Compile Include="MpvLib\MpvWrapper.cs" />
     <Compile Include="Classes\OperationInfo.cs" />
     <Compile Include="Classes\VideoFile.cs" />
-    <Compile Include="Classes\DedupEngine.cs" />
     <Compile Include="Classes\CodecInfo.cs" />
     <Compile Include="Enums\ProgressStyle.cs" />
     <Compile Include="Enums\ResolveOperation.cs" />
@@ -79,23 +67,6 @@
     <Compile Include="Interfaces\IVideoFile.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Interfaces\IFolderSettings.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\ImageComparison\ImageComparison.csproj">
-      <Project>{9317de4a-213d-4afc-b58f-ed401204534a}</Project>
-      <Name>ImageComparison</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="Libs\libmpv\x64\mpv-1.dll">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="Libs\libmpv\x86\mpv-1.dll">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
